# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: ISMS Builder API
  version: 1.0.0
  description: |
    REST API for the ISMS Builder platform — a self-hosted Information Security
    Management System (ISMS) according to ISO 27001:2022, NIS2, GDPR, BSI
    IT-Grundschutz and other frameworks.

    **Authentication:** All protected endpoints require a valid JWT cookie
    (`sm_session`). Obtain it via `POST /login`.

    **RBAC Roles (ascending rank):**
    - `reader` / `revision` (rank 1) — read access
    - `editor` / `dept_head` / `qmb` (rank 2) — create/edit
    - `contentowner` / `auditor` (rank 3) — approve, CISO/DSB settings
    - `admin` (rank 4) — user management, delete, system config

  contact:
    name: Claude Hecker
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Self-hosted instance
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sm_session
  schemas:
    CommonFields:
      type: object
      properties:
        id:
          type: string
          example: lx3k8fab2c
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        linkedControls:
          type: array
          items:
            type: string
          example: ["ISO-5.1", "BSI-ORP.3"]
        linkedPolicies:
          type: array
          items:
            type: string
...
    Template:
      allOf:
        - $ref: '#/components/schemas/CommonFields'
        - type: object
          properties:
            type:
              type: string
            language:
              type: string
            title:
              type: string
            content:
              type: string
            owner:
              type: string
            version:
              type: integer
            status:
              type: string
              enum: [draft, review, approved, archived]
            parentId:
              type: string
              nullable: true
            sortOrder:
              type: integer
            nextReviewDate:
              type: string
              format: date
            applicableEntities:
              type: array
              items:
                type: string
            attachments:
              type: array
              items:
                type: object
...
    Entity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        short:
          type: string
        type:
          type: string
          enum: [holding, subsidiary, branch, department]
        parentId:
          type: string
          nullable: true
        sortOrder:
          type: integer
...
paths:
...
  /templates:
    get:
      tags: [templates]
      summary: List templates
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Array of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
  /templates/tree:
    get:
      tags: [templates]
      summary: Get templates as a nested tree (by type/language)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Nested tree structure
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /templates/reorder:
    post:
      tags: [templates]
      summary: Batch-update sortOrder for sibling templates (contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                updates:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      sortOrder:
                        type: integer
      responses:
        '200':
          description: Reordered
  /template:
    post:
      tags: [templates]
      summary: Create a new template (contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, language, title]
              properties:
                type:
                  type: string
                language:
                  type: string
                title:
                  type: string
                content:
                  type: string
                parentId:
                  type: string
                  nullable: true
                linkedControls:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /template/{type}/{id}:
    get:
      tags: [templates]
      summary: Get a single template
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          description: Not found
    put:
      tags: [templates]
      summary: Update a template (editor+)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
      responses:
        '200':
          description: Updated template
    delete:
      tags: [templates]
      summary: Soft-delete a template (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /template/{type}/{id}/move:
    put:
      tags: [templates]
      summary: Move template to new parent (contentowner+)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parentId:
                  type: string
                  nullable: true
                sortOrder:
                  type: integer
      responses:
        '200':
          description: Moved
        '400':
          description: Circular reference detected
  /template/{type}/{id}/history:
    get:
      tags: [templates]
      summary: Get version history for a template
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Array of historical versions
  /template/{type}/{id}/status-history:
    get:
      tags: [templates]
      summary: Get lifecycle status history
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Array of status change events
  /template/{type}/{id}/attachments:
    post:
      tags: [templates]
      summary: Upload attachment to a template (editor+, multipart/form-data)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Attachment metadata
  /template/{type}/{id}/attachments/{attId}/file:
    get:
      tags: [templates]
      summary: Download attachment file
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: attId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File binary
  /template/{type}/{id}/attachments/{attId}:
    delete:
      tags: [templates]
      summary: Delete attachment (editor+)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: attId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /template/{type}/{id}/permanent:
    delete:
      tags: [templates]
      summary: Permanently delete template (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted
  /template/{type}/{id}/restore:
    post:
      tags: [templates]
      summary: Restore soft-deleted template (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restored

  # ─── ENTITIES ────────────────────────────────────────────────────────────
  /entities:
    get:
      tags: [entities]
      summary: List all entities
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
    post:
      tags: [entities]
      summary: Create entity (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
      responses:
        '201':
          description: Created entity
  /entities/tree:
    get:
      tags: [entities]
      summary: Get entities as nested tree
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Nested entity tree
  /entities/{id}:
    get:
      tags: [entities]
      summary: Get single entity
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entity
    put:
      tags: [entities]
      summary: Update entity (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
      responses:
        '200':
          description: Updated
    delete:
      tags: [entities]
      summary: Delete entity (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ─── SOA ─────────────────────────────────────────────────────────────────
...